Squash LedgerBackedIndexService, SpannedIndexService and MeteredReadOnlyLedger into ReadOnlyLedgerImpl
changelog_begin
changelog_end
