+++ ledger/participant-integration-api/src/main/scala/platform/apiserver/StandaloneIndexService.scala
@@ -3,8 +3,6 @@
 
 package com.daml.platform.apiserver
 
-import java.io.File
-
 import akka.stream.Materializer
 import com.daml.ledger.api.domain
 import com.daml.ledger.configuration.LedgerId
@@ -15,10 +13,12 @@ import com.daml.lf.data.Time.Timestamp
 import com.daml.lf.engine.{Engine, ValueEnricher}
 import com.daml.logging.LoggingContext
 import com.daml.metrics.Metrics
-import com.daml.platform.index.JdbcIndex
+import com.daml.platform.index.ReadOnlyLedgerBuilder
 import com.daml.platform.packages.InMemoryPackageStore
+import com.daml.platform.server.api.validation.ErrorFactories
 import com.daml.platform.store.{DbSupport, LfValueTranslationCache}
 
+import java.io.File
 import scala.concurrent.ExecutionContextExecutor
 
 object StandaloneIndexService {
@@ -71,28 +71,29 @@ object StandaloneIndexService {
         val packageStore = loadDamlPackages()
         preloadPackages(packageStore)
       })
-      indexService <- JdbcIndex
-        .owner(
-          dbSupport = dbSupport,
-          ledgerId = domain.LedgerId(ledgerId),
-          participantId = participantId,
-          eventsPageSize = config.eventsPageSize,
-          eventsProcessingParallelism = config.eventsProcessingParallelism,
-          acsIdPageSize = config.acsIdPageSize,
-          acsIdFetchingParallelism = config.acsIdFetchingParallelism,
-          acsContractFetchingParallelism = config.acsContractFetchingParallelism,
-          acsGlobalParallelism = config.acsGlobalParallelism,
-          acsIdQueueLimit = config.acsIdQueueLimit,
-          servicesExecutionContext = servicesExecutionContext,
-          metrics = metrics,
-          lfValueTranslationCache = lfValueTranslationCache,
-          enricher = valueEnricher,
-          maxContractStateCacheSize = config.maxContractStateCacheSize,
-          maxContractKeyStateCacheSize = config.maxContractKeyStateCacheSize,
-          maxTransactionsInMemoryFanOutBufferSize = config.maxTransactionsInMemoryFanOutBufferSize,
-          enableInMemoryFanOutForLedgerApi = config.enableInMemoryFanOutForLedgerApi,
-        )
-        .map(index => new SpannedIndexService(new TimedIndexService(index, metrics)))
+      indexService <- ReadOnlyLedgerBuilder(
+        dbSupport = dbSupport,
+        initialLedgerId = domain.LedgerId(ledgerId),
+        participantId = participantId,
+        eventsPageSize = config.eventsPageSize,
+        eventsProcessingParallelism = config.eventsProcessingParallelism,
+        acsIdPageSize = config.acsIdPageSize,
+        acsIdFetchingParallelism = config.acsIdFetchingParallelism,
+        acsContractFetchingParallelism = config.acsContractFetchingParallelism,
+        acsGlobalParallelism = config.acsGlobalParallelism,
+        acsIdQueueLimit = config.acsIdQueueLimit,
+        servicesExecutionContext = servicesExecutionContext,
+        metrics = metrics,
+        lfValueTranslationCache = lfValueTranslationCache,
+        enricher = valueEnricher,
+        maxContractStateCacheSize = config.maxContractStateCacheSize,
+        maxContractKeyStateCacheSize = config.maxContractKeyStateCacheSize,
+        maxTransactionsInMemoryFanOutBufferSize = config.maxTransactionsInMemoryFanOutBufferSize,
+        enableInMemoryFanOutForLedgerApi = config.enableInMemoryFanOutForLedgerApi,
+        errorFactories = ErrorFactories(),
+      )(materializer, loggingContext, servicesExecutionContext)
+        .owner()
+        .map(index => new TimedIndexService(index, metrics))
     } yield indexService
   }
 }
