--- ledger/participant-integration-api/src/main/scala/platform/index/MeteredReadOnlyLedger.scala
@@ -1,180 +1,0 @@
-// Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.daml.platform.index
-
-import akka.NotUsed
-import akka.stream.scaladsl.Source
-import com.daml.daml_lf_dev.DamlLf.Archive
-import com.daml.ledger.api.domain.{LedgerId, PartyDetails}
-import com.daml.ledger.api.health.HealthStatus
-import com.daml.ledger.api.v1.active_contracts_service.GetActiveContractsResponse
-import com.daml.ledger.api.v1.command_completion_service.CompletionStreamResponse
-import com.daml.ledger.api.v1.transaction_service.{
-  GetFlatTransactionResponse,
-  GetTransactionResponse,
-  GetTransactionTreesResponse,
-  GetTransactionsResponse,
-}
-import com.daml.ledger.configuration.Configuration
-import com.daml.ledger.offset.Offset
-import com.daml.ledger.participant.state.index.v2.MeteringStore.ReportData
-import com.daml.ledger.participant.state.index.v2.{MaximumLedgerTime, PackageDetails}
-import com.daml.lf.data.Ref
-import com.daml.lf.data.Ref.ApplicationId
-import com.daml.lf.data.Time.Timestamp
-import com.daml.lf.transaction.GlobalKey
-import com.daml.lf.value.Value.{ContractId, VersionedContractInstance}
-import com.daml.logging.LoggingContext
-import com.daml.metrics.{Metrics, Timed}
-import com.daml.platform.store.entries.{ConfigurationEntry, PackageLedgerEntry, PartyLedgerEntry}
-
-import scala.concurrent.Future
-
-private[index] class MeteredReadOnlyLedger(ledger: ReadOnlyLedger, metrics: Metrics)
-    extends ReadOnlyLedger {
-
-  override def ledgerId: LedgerId = ledger.ledgerId
-
-  override def currentHealth(): HealthStatus = ledger.currentHealth()
-
-  override def flatTransactions(
-      startExclusive: Option[Offset],
-      endInclusive: Option[Offset],
-      filter: Map[Ref.Party, Set[Ref.Identifier]],
-      verbose: Boolean,
-  )(implicit loggingContext: LoggingContext): Source[(Offset, GetTransactionsResponse), NotUsed] =
-    ledger.flatTransactions(startExclusive, endInclusive, filter, verbose)
-
-  override def transactionTrees(
-      startExclusive: Option[Offset],
-      endInclusive: Option[Offset],
-      requestingParties: Set[Ref.Party],
-      verbose: Boolean,
-  )(implicit
-      loggingContext: LoggingContext
-  ): Source[(Offset, GetTransactionTreesResponse), NotUsed] =
-    ledger.transactionTrees(startExclusive, endInclusive, requestingParties, verbose)
-
-  override def ledgerEnd()(implicit loggingContext: LoggingContext): Offset = ledger.ledgerEnd()
-
-  override def completions(
-      startExclusive: Option[Offset],
-      endInclusive: Option[Offset],
-      applicationId: Ref.ApplicationId,
-      parties: Set[Ref.Party],
-  )(implicit loggingContext: LoggingContext): Source[(Offset, CompletionStreamResponse), NotUsed] =
-    ledger.completions(startExclusive, endInclusive, applicationId, parties)
-
-  override def activeContracts(
-      filter: Map[Ref.Party, Set[Ref.Identifier]],
-      verbose: Boolean,
-  )(implicit
-      loggingContext: LoggingContext
-  ): (Source[GetActiveContractsResponse, NotUsed], Offset) =
-    ledger.activeContracts(filter, verbose)
-
-  override def lookupContract(
-      contractId: ContractId,
-      forParties: Set[Ref.Party],
-  )(implicit
-      loggingContext: LoggingContext
-  ): Future[Option[VersionedContractInstance]] =
-    Timed.future(metrics.daml.index.lookupContract, ledger.lookupContract(contractId, forParties))
-
-  override def lookupKey(key: GlobalKey, forParties: Set[Ref.Party])(implicit
-      loggingContext: LoggingContext
-  ): Future[Option[ContractId]] =
-    Timed.future(metrics.daml.index.lookupKey, ledger.lookupKey(key, forParties))
-
-  override def lookupFlatTransactionById(
-      transactionId: Ref.TransactionId,
-      requestingParties: Set[Ref.Party],
-  )(implicit loggingContext: LoggingContext): Future[Option[GetFlatTransactionResponse]] =
-    Timed.future(
-      metrics.daml.index.lookupFlatTransactionById,
-      ledger.lookupFlatTransactionById(transactionId, requestingParties),
-    )
-
-  override def lookupTransactionTreeById(
-      transactionId: Ref.TransactionId,
-      requestingParties: Set[Ref.Party],
-  )(implicit loggingContext: LoggingContext): Future[Option[GetTransactionResponse]] =
-    Timed.future(
-      metrics.daml.index.lookupTransactionTreeById,
-      ledger.lookupTransactionTreeById(transactionId, requestingParties),
-    )
-
-  override def lookupMaximumLedgerTime(
-      contractIds: Set[ContractId]
-  )(implicit loggingContext: LoggingContext): Future[MaximumLedgerTime] =
-    Timed.future(
-      metrics.daml.index.lookupMaximumLedgerTime,
-      ledger.lookupMaximumLedgerTime(contractIds),
-    )
-
-  override def getParties(parties: Seq[Ref.Party])(implicit
-      loggingContext: LoggingContext
-  ): Future[List[PartyDetails]] =
-    Timed.future(metrics.daml.index.getParties, ledger.getParties(parties))
-
-  override def listKnownParties()(implicit
-      loggingContext: LoggingContext
-  ): Future[List[PartyDetails]] =
-    Timed.future(metrics.daml.index.listKnownParties, ledger.listKnownParties())
-
-  override def partyEntries(startExclusive: Offset)(implicit
-      loggingContext: LoggingContext
-  ): Source[(Offset, PartyLedgerEntry), NotUsed] =
-    ledger.partyEntries(startExclusive)
-
-  override def listLfPackages()(implicit
-      loggingContext: LoggingContext
-  ): Future[Map[Ref.PackageId, PackageDetails]] =
-    Timed.future(metrics.daml.index.listLfPackages, ledger.listLfPackages())
-
-  override def getLfArchive(packageId: Ref.PackageId)(implicit
-      loggingContext: LoggingContext
-  ): Future[Option[Archive]] =
-    Timed.future(metrics.daml.index.getLfArchive, ledger.getLfArchive(packageId))
-
-  override def packageEntries(
-      startExclusive: Offset
-  )(implicit loggingContext: LoggingContext): Source[(Offset, PackageLedgerEntry), NotUsed] =
-    ledger.packageEntries(startExclusive)
-
-  override def lookupLedgerConfiguration()(implicit
-      loggingContext: LoggingContext
-  ): Future[Option[(Offset, Configuration)]] =
-    Timed.future(metrics.daml.index.lookupLedgerConfiguration, ledger.lookupLedgerConfiguration())
-
-  override def configurationEntries(
-      startExclusive: Offset
-  )(implicit loggingContext: LoggingContext): Source[(Offset, ConfigurationEntry), NotUsed] =
-    ledger.configurationEntries(startExclusive)
-
-  override def prune(
-      pruneUpToInclusive: Offset,
-      pruneAllDivulgedContracts: Boolean,
-  )(implicit loggingContext: LoggingContext): Future[Unit] =
-    Timed.future(
-      metrics.daml.index.prune,
-      ledger.prune(pruneUpToInclusive, pruneAllDivulgedContracts),
-    )
-
-  override def meteringReportData(
-      from: Timestamp,
-      to: Option[Timestamp],
-      applicationId: Option[ApplicationId],
-  )(implicit loggingContext: LoggingContext): Future[ReportData] = {
-    Timed.future(
-      metrics.daml.index.getTransactionMetering,
-      ledger.meteringReportData(from, to, applicationId),
-    )
-  }
-}
-
-private[platform] object MeteredReadOnlyLedger {
-  def apply(ledger: ReadOnlyLedger, metrics: Metrics): ReadOnlyLedger =
-    new MeteredReadOnlyLedger(ledger, metrics)
-}
