+++ ledger/participant-integration-api/src/main/scala/platform/index/ReadOnlyLedgerImpl.scala
@@ -6,8 +6,17 @@ package com.daml.platform.index
 import akka.NotUsed
 import akka.stream.scaladsl.Source
 import com.daml.daml_lf_dev.DamlLf
-import com.daml.ledger.api.domain
-import com.daml.ledger.api.domain.LedgerId
+import com.daml.error.DamlContextualizedErrorLogger
+import com.daml.ledger.api.{TraceIdentifiers, domain}
+import com.daml.ledger.api.domain.ConfigurationEntry.Accepted
+import com.daml.ledger.api.domain.{
+  LedgerId,
+  LedgerOffset,
+  PackageEntry,
+  PartyEntry,
+  TransactionFilter,
+  TransactionId,
+}
 import com.daml.ledger.api.health.HealthStatus
 import com.daml.ledger.api.v1.active_contracts_service.GetActiveContractsResponse
 import com.daml.ledger.api.v1.command_completion_service.CompletionStreamResponse
@@ -21,165 +30,373 @@ import com.daml.ledger.configuration.Configuration
 import com.daml.ledger.offset.Offset
 import com.daml.ledger.participant.state.index.v2
 import com.daml.ledger.participant.state.index.v2.MeteringStore.ReportData
-import com.daml.ledger.participant.state.index.v2.{ContractStore, MaximumLedgerTime}
+import com.daml.ledger.participant.state.index.v2.{
+  ContractStore,
+  IndexService,
+  MaximumLedgerTime,
+  _,
+}
 import com.daml.lf.data.Ref
+import com.daml.lf.data.Ref.{ApplicationId, Identifier, Party}
 import com.daml.lf.data.Time.Timestamp
 import com.daml.lf.transaction.GlobalKey
 import com.daml.lf.value.Value.{ContractId, VersionedContractInstance}
-import com.daml.logging.LoggingContext
-import com.daml.platform.PruneBuffers
+import com.daml.logging.{ContextualizedLogger, LoggingContext}
+import com.daml.metrics.{Metrics, Timed}
+import com.daml.platform.ApiOffset.ApiOffsetConverter
+import com.daml.platform.{ApiOffset, PruneBuffers}
 import com.daml.platform.akkastreams.dispatcher.Dispatcher
 import com.daml.platform.akkastreams.dispatcher.SubSource.RangeSource
+import com.daml.platform.server.api.validation.ErrorFactories
 import com.daml.platform.store.appendonlydao.{LedgerDaoTransactionsReader, LedgerReadDao}
-import com.daml.platform.store.entries.{ConfigurationEntry, PackageLedgerEntry, PartyLedgerEntry}
+import com.daml.platform.store.entries.PartyLedgerEntry
+import com.daml.telemetry.{Event, SpanAttribute, Spans}
+import scalaz.syntax.tag.ToTagOps
 
-import scala.concurrent.Future
+import scala.concurrent.{ExecutionContext, Future}
 
 private[index] class ReadOnlyLedgerImpl(
     val ledgerId: LedgerId,
+    participantId: Ref.ParticipantId,
     ledgerDao: LedgerReadDao,
     transactionsReader: LedgerDaoTransactionsReader,
     contractStore: ContractStore,
     pruneBuffers: PruneBuffers,
     dispatcher: Dispatcher[Offset],
-) extends ReadOnlyLedger {
+    metrics: Metrics,
+    errorFactories: ErrorFactories,
+) extends IndexService {
+  private val logger = ContextualizedLogger.get(getClass)
+  // TODO revisit redundant args lists
+  override def getLedgerId()(implicit loggingContext: LoggingContext): Future[LedgerId] =
+    Future.successful(ledgerId)
+
+  override def getParticipantId()(implicit
+      loggingContext: LoggingContext
+  ): Future[Ref.ParticipantId] =
+    Future.successful(participantId)
 
   override def currentHealth(): HealthStatus = ledgerDao.currentHealth()
 
-  override def lookupKey(key: GlobalKey, forParties: Set[Ref.Party])(implicit
+  override def lookupContractKey(readers: Set[Ref.Party], key: GlobalKey)(implicit
       loggingContext: LoggingContext
   ): Future[Option[ContractId]] =
-    contractStore.lookupContractKey(forParties, key)
+    Timed.future(metrics.daml.index.lookupKey, contractStore.lookupContractKey(readers, key))
 
-  override def flatTransactions(
-      startExclusive: Option[Offset],
-      endInclusive: Option[Offset],
-      filter: Map[Ref.Party, Set[Ref.Identifier]],
+  override def transactions(
+      startExclusive: domain.LedgerOffset,
+      endInclusive: Option[domain.LedgerOffset],
+      filter: domain.TransactionFilter,
       verbose: Boolean,
-  )(implicit loggingContext: LoggingContext): Source[(Offset, GetTransactionsResponse), NotUsed] =
-    dispatcher.startingAt(
-      startExclusive.getOrElse(Offset.beforeBegin),
-      RangeSource(transactionsReader.getFlatTransactions(_, _, filter, verbose)),
-      endInclusive,
+  )(implicit loggingContext: LoggingContext): Source[GetTransactionsResponse, NotUsed] =
+    between(startExclusive, endInclusive)((from, to) => {
+      from.foreach(offset =>
+        Spans.setCurrentSpanAttribute(SpanAttribute.OffsetFrom, offset.toHexString)
+      )
+      to.foreach(offset =>
+        Spans.setCurrentSpanAttribute(SpanAttribute.OffsetTo, offset.toHexString)
+      )
+      dispatcher
+        .startingAt(
+          from.getOrElse(Offset.beforeBegin),
+          RangeSource(transactionsReader.getFlatTransactions(_, _, convertFilter(filter), verbose)),
+          to,
+        )
+        .map(_._2)
+    }).wireTap(
+      _.transactions
+        .map(transaction =>
+          Event(transaction.commandId, TraceIdentifiers.fromTransaction(transaction))
+        )
+        .foreach(Spans.addEventToCurrentSpan)
     )
 
   override def transactionTrees(
-      startExclusive: Option[Offset],
-      endInclusive: Option[Offset],
-      requestingParties: Set[Ref.Party],
+      startExclusive: LedgerOffset,
+      endInclusive: Option[LedgerOffset],
+      filter: domain.TransactionFilter,
       verbose: Boolean,
-  )(implicit
-      loggingContext: LoggingContext
-  ): Source[(Offset, GetTransactionTreesResponse), NotUsed] =
-    dispatcher.startingAt(
-      startExclusive.getOrElse(Offset.beforeBegin),
-      RangeSource(
-        transactionsReader.getTransactionTrees(_, _, requestingParties, verbose)
-      ),
-      endInclusive,
+  )(implicit loggingContext: LoggingContext): Source[GetTransactionTreesResponse, NotUsed] =
+    between(startExclusive, endInclusive)((from, to) => {
+      from.foreach(offset =>
+        Spans.setCurrentSpanAttribute(SpanAttribute.OffsetFrom, offset.toHexString)
+      )
+      to.foreach(offset =>
+        Spans.setCurrentSpanAttribute(SpanAttribute.OffsetTo, offset.toHexString)
+      )
+      dispatcher
+        .startingAt(
+          from.getOrElse(Offset.beforeBegin),
+          RangeSource(
+            transactionsReader.getTransactionTrees(_, _, filter.filtersByParty.keySet, verbose)
+          ),
+          to,
+        )
+        .map(_._2)
+    }).wireTap(
+      _.transactions
+        .map(transaction =>
+          Event(transaction.commandId, TraceIdentifiers.fromTransactionTree(transaction))
+        )
+        .foreach(Spans.addEventToCurrentSpan)
     )
 
-  override def ledgerEnd()(implicit loggingContext: LoggingContext): Offset = dispatcher.getHead()
-
-  override def completions(
-      startExclusive: Option[Offset],
-      endInclusive: Option[Offset],
+  override def getCompletions(
+      startExclusive: LedgerOffset,
       applicationId: Ref.ApplicationId,
       parties: Set[Ref.Party],
-  )(implicit loggingContext: LoggingContext): Source[(Offset, CompletionStreamResponse), NotUsed] =
-    dispatcher.startingAt(
-      startExclusive.getOrElse(Offset.beforeBegin),
-      RangeSource(ledgerDao.completions.getCommandCompletions(_, _, applicationId, parties)),
-      endInclusive,
-    )
-
-  override def activeContracts(
-      filter: Map[Ref.Party, Set[Ref.Identifier]],
-      verbose: Boolean,
-  )(implicit
-      loggingContext: LoggingContext
-  ): (Source[GetActiveContractsResponse, NotUsed], Offset) = {
-    val activeAt = ledgerEnd()
-    (ledgerDao.transactionsReader.getActiveContracts(activeAt, filter, verbose), activeAt)
+  )(implicit loggingContext: LoggingContext): Source[CompletionStreamResponse, NotUsed] = {
+    convertOffset(startExclusive).flatMapConcat { beginOpt =>
+      dispatcher
+        .startingAt(
+          beginOpt,
+          RangeSource(ledgerDao.completions.getCommandCompletions(_, _, applicationId, parties)),
+          None,
+        )
+        .map(_._2)
+    }
   }
 
-  override def lookupContract(
-      contractId: ContractId,
+  override def getCompletions(
+      startExclusive: LedgerOffset,
+      endInclusive: LedgerOffset,
+      applicationId: Ref.ApplicationId,
+      parties: Set[Ref.Party],
+  )(implicit loggingContext: LoggingContext): Source[CompletionStreamResponse, NotUsed] =
+    between(startExclusive, Some(endInclusive))((start, end) =>
+      dispatcher
+        .startingAt(
+          start.getOrElse(Offset.beforeBegin),
+          RangeSource(ledgerDao.completions.getCommandCompletions(_, _, applicationId, parties)),
+          end,
+        )
+        .map(_._2)
+    )
+
+  override def getActiveContracts(
+      filter: TransactionFilter,
+      verbose: Boolean,
+  )(implicit loggingContext: LoggingContext): Source[GetActiveContractsResponse, NotUsed] = {
+    val currentLedgerEnd = ledgerEnd()
+    val acs =
+      ledgerDao.transactionsReader.getActiveContracts(
+        currentLedgerEnd,
+        convertFilter(filter),
+        verbose,
+      )
+    acs.concat(
+      Source.single(GetActiveContractsResponse(offset = ApiOffset.toApiString(currentLedgerEnd)))
+    )
+  }
+
+  override def lookupActiveContract(
       forParties: Set[Ref.Party],
+      contractId: ContractId,
   )(implicit
       loggingContext: LoggingContext
   ): Future[Option[VersionedContractInstance]] =
-    contractStore.lookupActiveContract(forParties, contractId)
+    Timed.future(
+      metrics.daml.index.lookupContract,
+      contractStore.lookupActiveContract(forParties, contractId),
+    )
 
-  override def lookupFlatTransactionById(
-      transactionId: Ref.TransactionId,
+  override def getTransactionById(
+      transactionId: TransactionId,
       requestingParties: Set[Ref.Party],
   )(implicit loggingContext: LoggingContext): Future[Option[GetFlatTransactionResponse]] =
-    ledgerDao.transactionsReader.lookupFlatTransactionById(transactionId, requestingParties)
+    Timed.future(
+      metrics.daml.index.lookupFlatTransactionById,
+      ledgerDao.transactionsReader
+        .lookupFlatTransactionById(transactionId.unwrap, requestingParties),
+    )
 
-  override def lookupTransactionTreeById(
-      transactionId: Ref.TransactionId,
+  override def getTransactionTreeById(
+      transactionId: TransactionId,
       requestingParties: Set[Ref.Party],
   )(implicit loggingContext: LoggingContext): Future[Option[GetTransactionResponse]] =
-    ledgerDao.transactionsReader.lookupTransactionTreeById(transactionId, requestingParties)
+    Timed.future(
+      metrics.daml.index.lookupTransactionTreeById,
+      ledgerDao.transactionsReader
+        .lookupTransactionTreeById(transactionId.unwrap, requestingParties),
+    )
 
-  override def lookupMaximumLedgerTime(
+  override def lookupMaximumLedgerTimeAfterInterpretation(
       contractIds: Set[ContractId]
   )(implicit loggingContext: LoggingContext): Future[MaximumLedgerTime] =
-    contractStore.lookupMaximumLedgerTimeAfterInterpretation(contractIds)
+    Timed.future(
+      metrics.daml.index.lookupMaximumLedgerTime,
+      contractStore.lookupMaximumLedgerTimeAfterInterpretation(contractIds),
+    )
 
   override def getParties(parties: Seq[Ref.Party])(implicit
       loggingContext: LoggingContext
   ): Future[List[domain.PartyDetails]] =
-    ledgerDao.getParties(parties)
+    Timed.future(metrics.daml.index.getParties, ledgerDao.getParties(parties))
 
   override def listKnownParties()(implicit
       loggingContext: LoggingContext
   ): Future[List[domain.PartyDetails]] =
-    ledgerDao.listKnownParties()
+    Timed.future(metrics.daml.index.listKnownParties, ledgerDao.listKnownParties())
 
-  override def partyEntries(startExclusive: Offset)(implicit
-      loggingContext: LoggingContext
-  ): Source[(Offset, PartyLedgerEntry), NotUsed] =
-    dispatcher.startingAt(startExclusive, RangeSource(ledgerDao.getPartyEntries))
+  override def partyEntries(
+      startExclusive: Option[LedgerOffset.Absolute]
+  )(implicit loggingContext: LoggingContext): Source[PartyEntry, NotUsed] = {
+    Source
+      .future(concreteOffset(startExclusive))
+      .flatMapConcat(dispatcher.startingAt(_, RangeSource(ledgerDao.getPartyEntries)))
+      .map {
+        case (_, PartyLedgerEntry.AllocationRejected(subId, _, reason)) =>
+          PartyEntry.AllocationRejected(subId, reason)
+        case (_, PartyLedgerEntry.AllocationAccepted(subId, _, details)) =>
+          PartyEntry.AllocationAccepted(subId, details)
+      }
+  }
 
   override def listLfPackages()(implicit
       loggingContext: LoggingContext
   ): Future[Map[Ref.PackageId, v2.PackageDetails]] =
-    ledgerDao.listLfPackages()
+    Timed.future(metrics.daml.index.listLfPackages, ledgerDao.listLfPackages())
 
   override def getLfArchive(packageId: Ref.PackageId)(implicit
       loggingContext: LoggingContext
   ): Future[Option[DamlLf.Archive]] =
-    ledgerDao.getLfArchive(packageId)
+    Timed.future(metrics.daml.index.getLfArchive, ledgerDao.getLfArchive(packageId))
 
-  override def packageEntries(startExclusive: Offset)(implicit
-      loggingContext: LoggingContext
-  ): Source[(Offset, PackageLedgerEntry), NotUsed] =
-    dispatcher.startingAt(startExclusive, RangeSource(ledgerDao.getPackageEntries))
+  override def packageEntries(
+      startExclusive: Option[LedgerOffset.Absolute]
+  )(implicit loggingContext: LoggingContext): Source[PackageEntry, NotUsed] =
+    Source
+      .future(concreteOffset(startExclusive))
+      .flatMapConcat(dispatcher.startingAt(_, RangeSource(ledgerDao.getPackageEntries)))
+      .map(_._2.toDomain)
 
-  override def lookupLedgerConfiguration()(implicit
+  /** Looks up the current configuration, if set, and the offset from which
+    * to subscribe to further configuration changes.
+    * The offset is internal and not exposed over Ledger API.
+    */
+  override def lookupConfiguration()(implicit
       loggingContext: LoggingContext
-  ): Future[Option[(Offset, Configuration)]] =
-    ledgerDao.lookupLedgerConfiguration()
+  ): Future[Option[(LedgerOffset.Absolute, Configuration)]] =
+    Timed.future(
+      metrics.daml.index.lookupLedgerConfiguration,
+      ledgerDao
+        .lookupLedgerConfiguration()
+        .map(
+          _.map { case (offset, config) => (toAbsolute(offset), config) }
+        )(ExecutionContext.parasitic),
+    )
 
-  override def configurationEntries(startExclusive: Offset)(implicit
+  /** Looks up the current configuration, if set, and continues to stream configuration changes.
+    */
+  override def getLedgerConfiguration()(implicit
       loggingContext: LoggingContext
-  ): Source[(Offset, ConfigurationEntry), NotUsed] =
-    dispatcher.startingAt(startExclusive, RangeSource(ledgerDao.getConfigurationEntries))
+  ): Source[LedgerConfiguration, NotUsed] = {
+    Source
+      .future(lookupConfiguration())
+      .flatMapConcat { optResult =>
+        val offset = optResult.map(_._1)
+        val foundConfig = optResult.map(_._2)
+
+        val initialConfig = Source(foundConfig.toList)
+        val configStream = configurationEntries(offset).collect {
+          case (_, Accepted(_, configuration)) => configuration
+        }
+        initialConfig
+          .concat(configStream)
+          .map(cfg => LedgerConfiguration(cfg.maxDeduplicationDuration))
+      }
+  }
+
+  /** Retrieve configuration entries. */
+  override def configurationEntries(startExclusive: Option[LedgerOffset.Absolute])(implicit
+      loggingContext: LoggingContext
+  ): Source[(domain.LedgerOffset.Absolute, domain.ConfigurationEntry), NotUsed] =
+    Source
+      .future(concreteOffset(startExclusive))
+      .flatMapConcat(dispatcher.startingAt(_, RangeSource(ledgerDao.getConfigurationEntries)).map {
+        case (offset, config) =>
+          toAbsolute(offset) -> config.toDomain
+      })
 
   override def prune(pruneUpToInclusive: Offset, pruneAllDivulgedContracts: Boolean)(implicit
       loggingContext: LoggingContext
-  ): Future[Unit] = {
-    pruneBuffers(pruneUpToInclusive)
-    ledgerDao.prune(pruneUpToInclusive, pruneAllDivulgedContracts)
-  }
+  ): Future[Unit] = Timed.future(
+    metrics.daml.index.prune, {
+      pruneBuffers(pruneUpToInclusive)
+      ledgerDao.prune(pruneUpToInclusive, pruneAllDivulgedContracts)
+    },
+  )
 
-  override def meteringReportData(
+  override def getMeteringReportData(
       from: Timestamp,
       to: Option[Timestamp],
-      applicationId: Option[Ref.ApplicationId],
-  )(implicit loggingContext: LoggingContext): Future[ReportData] = {
-    ledgerDao.meteringReportData(from, to, applicationId)
+      applicationId: Option[ApplicationId],
+  )(implicit loggingContext: LoggingContext): Future[ReportData] =
+    Timed.future(
+      metrics.daml.index.getTransactionMetering,
+      ledgerDao.meteringReportData(
+        from: Timestamp,
+        to: Option[Timestamp],
+        applicationId: Option[ApplicationId],
+      ),
+    )
+
+  override def currentLedgerEnd()(implicit
+      loggingContext: LoggingContext
+  ): Future[LedgerOffset.Absolute] = {
+    val offset =
+      if (ledgerEnd() == Offset.beforeBegin) ApiOffset.begin
+      else ledgerEnd()
+    Future.successful(toAbsolute(offset))
   }
+
+  private def ledgerEnd(): Offset = dispatcher.getHead()
+
+  // Returns a function that memoizes the current end
+  // Can be used directly or shared throughout a request processing
+  private def convertOffset: LedgerOffset => Source[Offset, NotUsed] = {
+    case LedgerOffset.LedgerBegin => Source.single(Offset.beforeBegin)
+    case LedgerOffset.LedgerEnd => Source.single(ledgerEnd())
+    case LedgerOffset.Absolute(offset) =>
+      ApiOffset.fromString(offset).fold(Source.failed, off => Source.single(off))
+  }
+
+  private def between[A](
+      startExclusive: domain.LedgerOffset,
+      endInclusive: Option[domain.LedgerOffset],
+  )(f: (Option[Offset], Option[Offset]) => Source[A, NotUsed])(implicit
+      loggingContext: LoggingContext
+  ): Source[A, NotUsed] = {
+    val convert = convertOffset
+    convert(startExclusive).flatMapConcat { begin =>
+      endInclusive
+        .map(convert(_).map(Some(_)))
+        .getOrElse(Source.single(None))
+        .flatMapConcat {
+          case Some(`begin`) =>
+            Source.empty
+          case Some(end) if begin > end =>
+            Source.failed(
+              errorFactories.offsetOutOfRange(
+                s"End offset ${end.toApiString} is before Begin offset ${begin.toApiString}."
+              )(new DamlContextualizedErrorLogger(logger, loggingContext, None))
+            )
+          case endOpt: Option[Offset] =>
+            f(Some(begin), endOpt)
+        }
+    }
+  }
+
+  private def convertFilter(filter: TransactionFilter): Map[Party, Set[Identifier]] =
+    filter.filtersByParty.map { case (party, filters) =>
+      party -> filters.inclusive.fold(Set.empty[Identifier])(_.templateIds)
+    }
+
+  private def concreteOffset(startExclusive: Option[LedgerOffset.Absolute]): Future[Offset] =
+    startExclusive
+      .map(off => Future.fromTry(ApiOffset.fromString(off.value)))
+      .getOrElse(Future.successful(Offset.beforeBegin))
+
+  private def toAbsolute(offset: Offset): LedgerOffset.Absolute =
+    LedgerOffset.Absolute(offset.toApiString)
 }
